+++ apps/docs/components/icons/bold/index.ts
@@ -1,0 +1,6 @@
+export * from "./keyboard";
+export * from "./mouse-circle";
+export * from "./squares";
+export * from "./fatrows";
+export * from "./eye";
+export * from "./keyboard-open";