+++ apps/docs/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./nextui-logo";
-export * from "./docs-navbar";
 export * from "./theme-switch";
 export * from "./hero";
 export * from "./looper-bg";