+++ apps/docs/components/primitives.ts
@@ -17,8 +17,8 @@ export const title = tv({
     },
     size: {
       sm: "text-2xl lg:text-3xl",
-      md: "text-4xl lg:text-5xl",
-      lg: "text-5xl lg:text-6xl",
+      md: "text-3xl lg:text-5xl",
+      lg: "text-4xl lg:text-6xl",
     },
     fullWidth: {
       true: "w-full block",
@@ -36,7 +36,7 @@ export const title = tv({
 });
 
 export const subtitle = tv({
-  base: "w-full md:w-1/2 my-2 text-xl font-light text-neutral-500 block max-w-full",
+  base: "w-full md:w-1/2 my-2 text-lg lg:text-xl font-light text-neutral-500 block max-w-full",
   variants: {
     fullWidth: {
       true: "!w-full",