+++ apps/docs/libs/constants.ts
@@ -1,0 +1,2 @@
+export const TWITTER_USER_NAME = "getnextui";
+export const SITE_URL = "https://nextui.org";