+++ packages/components/tooltip/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nextui-org/tooltip
 
+## 0.0.0-dev-v2-20230501173002
+
+### Patch Changes
+
+- Popover overflow fixed
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230501173002
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230501173002
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230501173002
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230501173002
+  - @nextui-org/system@0.0.0-dev-v2-20230501173002
+  - @nextui-org/theme@0.0.0-dev-v2-20230501173002
+
 ## 0.0.0-dev-v2-20230429213333
 
 ### Patch Changes